chore: rename AppStack index
