+++ navigation/index.tsx
@@ -8,7 +8,7 @@ import { useState, useEffect } from 'react'
 import { ColorSchemeName } from 'react-native'
 
 import { AuthScreenStack } from '../screens/AuthStack'
-import { HomeScreenStack } from '../screens/AppStack/AppStack'
+import { HomeScreenStack } from '../screens/AppStack'
 import { supabase } from '../supabase-service'
 
 export default function Navigation({
